Remove db layer hard-code permission checks for quota_destroy_all_*
This patch removes the hard-code permission checks for db call
quota_destroy_all_by_project_and_user and quota_destroy_all_by_project.

For v2 API adds back-compatible checking. For v2.1 API, remove useless
project owner permission checks that should be done by policy.

And correct related unittest. For v2.1 API, the unittest should use
non-admin context. For v2 API, the unitest keep as before and add
non-admin context test to ensure the v2 back-comaptible.

Partially implements bp nova-api-policy-final-part

+++ nova/api/openstack/compute/contrib/quotas.py
@@ -220,6 +220,12 @@ class QuotaSetsController(wsgi.Controller):
                 raise webob.exc.HTTPNotFound()
             try:
                 nova.context.authorize_project_context(context, id)
+                # NOTE(alex_xu): back-compatible with db layer hard-code admin
+                # permission checks. This has to be left only for API v2.0
+                # because this version has to be stable even if it means that
+                # only admins can call this method while the policy could be
+                # changed.
+                nova.context.require_admin_context(context)
                 if user_id:
                     QUOTAS.destroy_all_by_project_and_user(context,
                                                            id, user_id)
